+++ bugsnag-android-core/src/main/java/com/bugsnag/android/EventStore.java
@@ -26,7 +26,6 @@ class EventStore extends FileStore {
     private static final String STARTUP_CRASH = "_startupcrash";
     private static final long LAUNCH_CRASH_TIMEOUT_MS = 2000;
     private static final int LAUNCH_CRASH_POLL_MS = 50;
-    private static final int MAX_EVENT_COUNT = 32;
 
     volatile boolean flushOnLaunchCompleted = false;
     private final Semaphore semaphore = new Semaphore(1);
@@ -54,9 +53,12 @@ class EventStore extends FileStore {
     };
 
     EventStore(@NonNull ImmutableConfig config,
-               @NonNull Context appContext, @NonNull Logger logger,
-               Notifier notifier, Delegate delegate) {
-        super(appContext, "/bugsnag-errors/", MAX_EVENT_COUNT, EVENT_COMPARATOR, logger, delegate);
+               @NonNull Context appContext,
+               @NonNull Logger logger,
+               Notifier notifier,
+               Delegate delegate) {
+        super(appContext, "/bugsnag-errors/", config.getMaxPersistedEvents(),
+                EVENT_COMPARATOR, logger, delegate);
         this.config = config;
         this.logger = logger;
         this.delegate = delegate;